# nova-proto/gen/openapiv2/nova/post/v1/post.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Post API",
    "version": "1.0",
    "contact": {
      "name": "nova project",
      "url": "https://github.com/miiy/goc-quickstart",
      "email": "none@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/miiy/goc-quickstart/LICENSE"
    }
  },
  "tags": [
    {
      "name": "PostService"
    }
  ],
  "schemes": [
    "http",
    "https",
    "wss"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/posts": {
      "get": {
        "summary": "ListPosts returns posts matched by filters and pagination.",
        "operationId": "PostService_ListPosts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPostsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "authorId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "PostService"
        ]
      },
      "post": {
        "summary": "CreatePost creates a new post.",
        "operationId": "PostService_CreatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreatePostResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "CreatePostRequest carries the post to create.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreatePostRequest"
            }
          }
        ],
        "tags": [
          "PostService"
        ]
      }
    },
    "/api/v1/posts/{id}": {
      "get": {
        "summary": "GetPost returns a single post by id.",
        "operationId": "PostService_GetPost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPostResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "PostService"
        ]
      },
      "delete": {
        "summary": "DeletePost deletes a post by id.",
        "operationId": "PostService_DeletePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeletePostResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "PostService"
        ]
      },
      "put": {
        "summary": "UpdatePost updates mutable fields for a post.",
        "operationId": "PostService_UpdatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdatePostResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostServiceUpdatePostBody"
            }
          }
        ],
        "tags": [
          "PostService"
        ]
      }
    }
  },
  "definitions": {
    "PostServiceUpdatePostBody": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        },
        "updateMask": {
          "type": "string"
        }
      },
      "description": "UpdatePostRequest carries post fields to update.",
      "required": [
        "post"
      ]
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreatePostRequest": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        }
      },
      "description": "CreatePostRequest carries the post to create.",
      "required": [
        "post"
      ]
    },
    "v1CreatePostResponse": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        }
      },
      "description": "CreatePostResponse returns the created post."
    },
    "v1DeletePostResponse": {
      "type": "object",
      "description": "DeletePostResponse is returned after a post is deleted."
    },
    "v1GetPostResponse": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        }
      },
      "description": "GetPostResponse returns a single post."
    },
    "v1ListPostsResponse": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "currentPage": {
          "type": "integer",
          "format": "int32"
        },
        "posts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Post"
          }
        }
      },
      "description": "ListPostsResponse returns a paginated list of posts."
    },
    "v1Post": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "authorId": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/v1PostStatus"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "categoryId": {
          "type": "string",
          "format": "int64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "deletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "authorName": {
          "type": "string",
          "description": "Populated by the API gateway from nova-user for web responses."
        }
      },
      "description": "Post describes an article published by a user."
    },
    "v1PostStatus": {
      "type": "string",
      "enum": [
        "POST_STATUS_UNSPECIFIED",
        "POST_STATUS_DRAFT",
        "POST_STATUS_PUBLISHED"
      ],
      "default": "POST_STATUS_UNSPECIFIED",
      "description": "PostStatus describes the publication state of a post."
    },
    "v1UpdatePostResponse": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        }
      },
      "description": "UpdatePostResponse returns the updated post."
    }
  },
  "externalDocs": {
    "description": "More about goc",
    "url": "https://github.com/grpc-ecosystem/grpc-gateway"
  }
}